API reference

BootstrapAsymptotics.BootstrapAsymptotics — Module

BootstrapAsymptotics

State evolution and simulation for bootstrap-related methods applied to linear and logistic regression.

Exports

BayesOpt

ERM

FullResampling

LabelResampling

Logistic

Overlaps

PairBootstrap

ResidualBootstrap

Ridge

Subsampling

bias_state_evolution

bias_variance_empirical

bias_variance_true

fit

gamp

sample_all

sample_data

sample_labels

sample_weights

state_evolution

state_evolution_BayesOpt

variance_state_evolution

BootstrapAsymptotics.BayesOpt — Type

struct BayesOpt <: BootstrapAsymptotics.Algorithm

Bayes optimal estimation algorithm.

BootstrapAsymptotics.ERM — Type

struct ERM <: BootstrapAsymptotics.Algorithm

Empirical Risk Minimization algorithm.

BootstrapAsymptotics.FullResampling — Type

struct FullResampling <: BootstrapAsymptotics.Algorithm

Full resampling algorithm.

BootstrapAsymptotics.LabelResampling — Type

struct LabelResampling <: BootstrapAsymptotics.Algorithm

Label resampling algorithm.

BootstrapAsymptotics.Logistic — Type

struct Logistic <: BootstrapAsymptotics.Problem

Logistic regression problem with ridge penalty.

Fields

α::Float64: ratio of population over dimension n/d

λ::Float64: regularization strength

ρ::Float64: teacher weight

BootstrapAsymptotics.Overlaps — Type

struct Overlaps{hat, T1<:(AbstractVector), T2<:(AbstractMatrix), T3<:(AbstractMatrix)}

Overlap or hat overlap storage for state evolution between two algorithms.

Fields

m::AbstractVector

Q::AbstractMatrix

V::AbstractMatrix

BootstrapAsymptotics.PairBootstrap — Type

struct PairBootstrap <: BootstrapAsymptotics.Algorithm

Standard (pair) bootstrap algorithm.

Fields

p_max::Int64: maximum weight for state evolution

BootstrapAsymptotics.ResidualBootstrap — Type

struct ResidualBootstrap <: BootstrapAsymptotics.Algorithm

Residual bootstrap algorithm, aka ERM + label resampling.

BootstrapAsymptotics.Ridge — Type

struct Ridge <: BootstrapAsymptotics.Problem

Least squares regression problem with ridge penalty.

Fields

α::Float64: ratio of population over dimension n/d

Δ::Float64: Gaussian noise variance

λ::Float64: regularization strength

ρ::Float64: teacher weight

BootstrapAsymptotics.Subsampling — Type

struct Subsampling <: BootstrapAsymptotics.Algorithm

Subsampling algorithm.

Fields

r::Float64: subsampling fraction

BootstrapAsymptotics.bias_state_evolution — Method

bias_state_evolution(
    problem,
    algo;
    check_convergence,
    kwargs...
)

BootstrapAsymptotics.bias_variance_empirical — Method

bias_variance_empirical(rng, problem, algo; n, K)

BootstrapAsymptotics.bias_variance_empirical — Method

bias_variance_empirical(rng, problem, algo; n, K)

BootstrapAsymptotics.bias_variance_empirical — Method

bias_variance_empirical(rng, problem, algo; n, K)

BootstrapAsymptotics.bias_variance_empirical — Method

bias_variance_empirical(rng, problem, algo; n, K)

BootstrapAsymptotics.bias_variance_true — Method

bias_variance_true(rng, problem; n, K, conditional)

BootstrapAsymptotics.gamp — Method

gamp(problem, X, y; max_iter, rtol)

Run generalized approximate message passing on logistic regression to recover the Bayes optimal estimator.

BootstrapAsymptotics.sample_all — Method

sample_all(rng, problem, n)

Sample X, w and y all at once for a given problem with population size n.

BootstrapAsymptotics.sample_data — Method

sample_data(rng, problem, n)

Sample the data matrix X for a given problem with population size n.

BootstrapAsymptotics.sample_labels — Method

sample_labels(rng, , X, w)

Sample the labels vector y for a given problem from the features X and weights w.

BootstrapAsymptotics.sample_weights — Method

sample_weights(rng, problem, n)

Sample the weights vector w for a given problem with population size n (from which the dimension is deduced).

BootstrapAsymptotics.state_evolution — Method

state_evolution(
    problem,
    algo1,
    algo2;
    rtol,
    max_iteration,
    show_progress
)

Peform state evolution on a problem for the couple (algorithm1, algorithm2), by creating and then iteratively updating overlaps and hat overlaps.

Keyword arguments

rtol: relative tolerance used at every step of the procedure, especially to check overlap convergence

max_iteration: maximum number of overlap updates

show_progress: whether to display a progress bar

BootstrapAsymptotics.state_evolution_BayesOpt — Method

state_evolution_BayesOpt(problem; rtol, max_iteration)

Special case of state evolution for the Bayes optimal estimator.

BootstrapAsymptotics.variance_state_evolution — Method

variance_state_evolution(
    problem,
    algo;
    check_convergence,
    kwargs...
)

BootstrapAsymptotics.variance_state_evolution — Method

variance_state_evolution(
    problem,
    algo;
    check_convergence,
    kwargs...
)

StatsAPI.fit — Function

fit([rng], problem, algorithm, X, y, [w_star])

Fit an algorithm to data X, y generated by problem, with randomness source rng.